+++ LTEV2X/VUE.h
@@ -20,11 +20,11 @@ public:
 	~cVeUE();
 	void initializeUrban(sUEConfigure &t_UEConfigure);
 	void initializeHigh(sUEConfigure &t_UEConfigure);
-	unsigned short m_wRoadID;
+	int m_wRoadID;
 	int m_locationID;
 	const int m_VeUEId = m_VeUECount++;
-	unsigned short m_RSUId;//RRM_DRA模块需要
-	unsigned short m_ClusterIdx;//RRM_DRA模块需要
+	int m_RSUId;//RRM_DRA模块需要
+	int m_ClusterIdx;//RRM_DRA模块需要
 	double m_fX;
 	double m_fY;
 	double m_fAbsX;
@@ -39,7 +39,7 @@ public:
 	int m_Nr;//接收天线数目
 	double m_Ploss;//路径损耗
 	double *m_H;//信道响应矩阵
-	unsigned short m_InterVeUENum;//同频干扰数量
+	int m_InterVeUENum;//同频干扰数量
 	std::vector<int> m_InterVeUEVec;//同频干扰车辆ID，不包含当前车辆
 	std::vector<double> m_InterferencePloss;//干扰路径损耗
 	double *m_InterferenceH = nullptr;//干扰信道响应矩阵
@@ -48,7 +48,7 @@ public:
 
 
 	/* HighSpeed */
-	unsigned short m_wLaneID;
+	int m_wLaneID;
 
 
 	std::list<std::tuple<int, int>> m_LocationUpdateLogInfoList;//地理位置更新日志信息
